Add Script to Test Binaries and Boards
Add a script that can be used to run a given binary, with a given board as input and return the average time it took to run. The binary, runs, and board can be configured via CLI flags.
This should help us compare the different implementations more easily, even with very small run times

Signed-off-by: Matthew Hill <GeistInDerSH@gmail.com>