Add Optional unchecked_indexing Feature
The generated ASM contains a few bounds check calls. Overall, these are not a big deal (testing on the remote machine showed between 0-15μs for the whole run). It is slightly more stable of a runtime though.
Add the option, but do not enable it by default, outside an additional solve_ feature

Signed-off-by: Matthew Hill <GeistInDerSH@gmail.com>