Build a single binary per solver
Add another entrypoint, with make target support, for building a separate binary for each of the sudoku solvers in main.c. This may allow for the compiler to better optimize some of the code for better runtime results.
Note that this version also includes the time to parse the file, unlike main.c so a pure comparison of the two is not one to one.

Signed-off-by: Matthew Hill <GeistInDerSH@gmail.com>