Add Full Run Benchmarks & Format Output
Add benchmark timings that includes reading from the file system. This helps to test how long it takes to detect an invalid file/board.

In addition, format the output so that all timing data has the same indentation to make it easier to quickly parse the output. The side effect of this is that the logs of invalid are no longer printed.

Signed-off-by: Matthew Hill <GeistInDerSH@gmail.com>