Add skip_intermediate_stack_frames Feature
Add an optional feature for the bit_masking v1/v2 implementation based of reading through C/sudoku_optimized_v2.c. There George calculates the next row/column that the solver should look at before recursively calling the solve function. This helps to avoid cases where we are simply incrementing off the end of the row. It saves a few stack frames per call which isn't a huge savin, but it is noticeable

Signed-off-by: Matthew Hill <GeistInDerSH@gmail.com>